Fully translated and tested Dialog_Trees.xml
I've fully translated Dialog_Trees.xml this file seems to be related to
the comms console and some, not all, options included in the comms
console regarding relations with other factions and some of the dialog
between those interactions through the console.
This has been fully tested in-game and translated without any in-game
errors or overlapping.